Fix: enhance access control for unauthorized users

+++ brand-partner/delete-product.php
@@ -1,5 +1,6 @@
 <?php
 include("../config.php");
+require("../functions/general.php");
 
 session_start();
 
@@ -17,6 +18,12 @@ if ($_SESSION["role"] == "partner") {
 
         $query = mysqli_query($conn, "SELECT * FROM products WHERE id = $id");
         $data = mysqli_fetch_assoc($query);
+
+        if($data['partner_id'] != $_SESSION['id']){
+            $_SESSION['alert'] = flash('Ouch!', 'You do not have permission to delete the product.', 'warning');
+            header("location: /NeoMall/brand-partner/manage-product.php");
+            exit();
+        }
 
         unlink($data['picture']);
